Optimize visibility and documentation
