+++ src/Psl/Type/Internal/ConvertedType.php
@@ -8,7 +8,6 @@ use Closure;
 use Psl\Type;
 use Psl\Type\Exception\AssertException;
 use Psl\Type\Exception\CoercionException;
-use Psl\Type\Exception\PathExpression;
 use Psl\Type\TypeInterface;
 use Throwable;
