+++ src/Psl/Type/Internal/MapType.php
@@ -10,7 +10,6 @@ use Psl\Str;
 use Psl\Type;
 use Psl\Type\Exception\AssertException;
 use Psl\Type\Exception\CoercionException;
-use Psl\Type\Exception\PathExpression;
 use Throwable;
 
 use function is_iterable;
