+++ src/Psl/Type/Internal/ShapeType.php
@@ -8,7 +8,6 @@ use Psl\Iter;
 use Psl\Type;
 use Psl\Type\Exception\AssertException;
 use Psl\Type\Exception\CoercionException;
-use Psl\Type\Exception\PathExpression;
 use stdClass;
 use Throwable;
