+++ src/Psl/Type/Internal/VecType.php
@@ -7,7 +7,6 @@ namespace Psl\Type\Internal;
 use Psl\Type;
 use Psl\Type\Exception\AssertException;
 use Psl\Type\Exception\CoercionException;
-use Psl\Type\Exception\PathExpression;
 use Throwable;
 
 use function array_is_list;
